moved styles into own file
